Added new Reading section.
If readability is so important to python, why don't python guides have a
list of recommended awesome code to read? Why don't they explain that
code? The idea behind this reading section is to provide such a guide.
The idea here is to start with something small and useful, with a vision
for the future, and iterate quickly.