Suggest improved introduction to Style/#Idioms
My reading, though limited, suggests that the general notion of Idioms
is used often, but rarely explained. It feels very alien to beginners. A
term like "Idiomatic Python" can cause mental stack overflow. It did to
me :) So I humbly submit, in this case, explicit is better than
implicit. I debated whether to place it at the top of the document,
where idioms are first mentioned, but placed it here instead, as this
section explicitly references the topic of idioms.