Intro paras for Freezing and Packaging Your Code
I've tried to describe why you would choose one method over
another, and what the relative advantages and disadvantages
are of each method.

I'm painfully aware that I haven't covered the combo of
freezing an application *then* creating a Linux distro
package of it, but maybe that's a battle for another day.

My intent is to add more content on the Freezing page
once this commit is accepted, so I can hear your feedback
before doing any further work.

Thanks.