(osx) how to make `python` reference `python3`
I noticed that the very next page (which is no longer OSX specific) makes the assumption that `python` will invoke the Python 3 interpreter, while the last mention of versions on this page leaves `python` referencing the OSX System Python interpreter.

I honestly do not know if this is the "correct" way to configure this default. It looks like brew is cautious about symlinking over the native Python 2.7.10 (High Sierra) and they might have good reason. All of my projects are Python 3 now, so i'm ready to "make it ~facebook~ PATH official with Python 3"

Regardless of how it's done - or if it's a bad idea - it's worth mentioning IMO because there is a discontinuity of where this page leaves `python` and where the next page assumes `python` to be.