Update diveinto link
The diveinto.org link is broken when I try it.  I found that https://diveinto.org/python3/table-of-contents.html works but the https://diveintopython3.net/ is shorter and appears less broken to someone reading the url for the first time.  The table of content at this updated address looks the same as what I found on the web.archive.org for the diveinto.org link.